misc/HtmlTestSuite/results/test-0237.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><link href="mailto:jarhar@chromium.org" rel="author"/>
<link href="http://crbug.com/1247417" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>

</head><body><div id="container1" style="content-visibility:hidden"/>
<div id="container2" style="content-visibility:hidden"/>

<script>
promise_test(async () =&gt; {
  const image1 = document.createElement('input');
  image1.type = 'image';
  const image1Load = new Promise(resolve =&gt; {
    image1.addEventListener('load', resolve);
  });
  image1.src = 'resources/dice.png';
  container1.appendChild(image1);
  await image1Load;
  assert_not_equals(image1.width, 0, 'width');

  const image2 = document.createElement('input');
  image2.type = 'image';
  const image2Load = new Promise(resolve =&gt; {
    image2.addEventListener('load', resolve);
  });
  image2.src = 'resources/dice.png';
  container2.appendChild(image2);
  await image2Load;
  assert_not_equals(image2.height, 0, 'height');

}, `&lt;input type=image&gt; should return nonzero values for width and height in a c-v:hidden subtree.`);
</script>
</body></html>
